+++ hutool-aop/src/main/java/cn/hutool/aop/interceptor/CglibInterceptor.java
@@ -41,7 +41,8 @@ public class CglibInterceptor implements MethodInterceptor, Serializable {
 		// 开始前回调
 		if (aspect.before(target, method, args)) {
 			try {
-				result = proxy.invokeSuper(obj, args);
+//				result = proxy.invokeSuper(obj, args);
+				result = proxy.invoke(target, args);
 			} catch (InvocationTargetException e) {
 				// 异常回调（只捕获业务代码导致的异常，而非反射导致的异常）
 				if (aspect.afterException(target, method, args, e.getTargetException())) {